moved propTypes and defaultProps for the RenderContext to `param-types`
