+++ app/client-app/src/app/pages/Editor/components/PagePreview.js
@@ -9,7 +9,7 @@ import {
   PageItemActionsProvider,
 } from "../contexts/PageItemActions";
 import pageItemActions from "../actions/pageItemActions";
-import { defaultProps as renderActions } from "constants/param-types";
+import { defaultProps as renderActions } from "@decsys/param-types/ResponseItemRenderContext";
 
 // TODO: Document this capability for components
 // with the api (props signature, how renderedItem works..., _context...)
