+++ app/client-app/src/app/pages/Editor/contexts/PageItemActions.js
@@ -1,9 +1,10 @@
 import { createContext, useContext } from "react";
 
 const PageItemActionsContext = createContext({
-  duplicatePageItem: itemId => {},
-  deletePageItem: itemId => {},
-  setParamValue: (itemId, paramKey, paramValue) => {}
+  duplicatePageItem: (itemId) => {},
+  deletePageItem: (itemId) => {},
+  setParamValue: (itemId, paramKey, paramValue) => {},
+  changePageResponseItem: (itemId, type, order) => {},
 });
 
 export const usePageItemActions = () => useContext(PageItemActionsContext);
